Score working

+++ controllers/business.js
@@ -2,11 +2,17 @@ const { matchedData } = require("express-validator")
 const Business = require("../models/business.js")
 const User = require("../models/user.js")
 const { handleHTTPError } = require("../utils/handleHTTPError.js")
+const { verifyToken } = require("../utils/handleJWT.js")
 
 const editBusiness = async (req, res) => { 
     try {
         const id = matchedData(req, { locations: ["query"] }).id
         const { activity, title, summary, images, texts} = matchedData(req, { locations: ["body"] })
+        const token = matchedData(req, { locations: ["headers"] }).authorization
+
+        const businessTokenData = verifyToken(token)
+
+        if(id !== businessTokenData._id) throw new Error("ERROR_BUSINESS_EDIT_NOT_PERMITED")
 
         const updatedBusiness = await Business.findOneAndUpdate(
             {_id: id}, 
@@ -22,20 +28,25 @@ const editBusiness = async (req, res) => {
         res.send(updatedBusiness)
     } catch(err) {
         // SLACK log I think
-        handleHTTPError(res, "ERROR_EDIT_BUSINESS")
+        handleHTTPError(res, err.message)
     }
 }
 
 const deleteBusiness = async (req, res) => {
     try{
         const id = matchedData(req, { locations: ["query"] }).id
+        const token = matchedData(req, { locations: ["headers"] }).authorization
+
+        const businessTokenData = verifyToken(token)
+
+        if(id !== businessTokenData._id) throw new Error("ERROR_BUSINESS_DELETE_NOT_PERMITED")
 
         const deletedBusiness = await Business.findOneAndDelete({_id:id})
 
         res.send(deletedBusiness)
     } catch(err){
         // SLACK log I think
-        handleHTTPError(res, "ERROR_DELETE_BUSINESS")
+        handleHTTPError(res, err.message)
     }
 }
 
@@ -43,6 +54,11 @@ const editImages = async (req, res) => {
     try {
         const id = matchedData(req, { locations: ["query"] }).id
         const newImages = matchedData(req, { locations: ["body"] }).images
+        const token = matchedData(req, { locations: ["headers"] }).authorization
+
+        const businessTokenData = verifyToken(token)
+
+        if(id !== businessTokenData._id) throw new Error("ERROR_BUSINESS_IMAGE_EDIT_NOT_PERMITED")
 
         const preBusiness = await Business.findById(id)
 
@@ -58,7 +74,7 @@ const editImages = async (req, res) => {
         res.send(updatedBusiness)
     } catch(err) {
         // SLACK log I think
-        handleHTTPError(res, "ERROR_EDIT_IMAGES")
+        handleHTTPError(res, err.mesage)
     }
 }
 
@@ -66,8 +82,14 @@ const editTexts = async (req, res) => {
     try {
         const id = matchedData(req, { locations: ["query"] }).id
         const newTexts = matchedData(req, { locations: ["body"] }).texts
+        const token = matchedData(req, { locations: ["headers"] }).authorization
+
+        const businessTokenData = verifyToken(token)
+
+        if(id !== businessTokenData._id) throw new Error("ERROR_BUSINESS_TEXTS_EDIT_NOT_PERMITED")
         
         const preBusiness = await Business.findById(id)
+
         let texts = preBusiness.texts
         texts.push(...newTexts)
 
@@ -79,23 +101,26 @@ const editTexts = async (req, res) => {
         res.send(updatedBusiness)
     } catch(err) {
         // SLACK log I think
-        handleHTTPError(res, "ERROR_EDIT_TEXTS")
+        handleHTTPError(res, err.message)
     }
 }
 
 // By providing the businesses id, it will search for all users in the same city than the business with the provided activity/interest and want to receive emails
 const getMailingList = async (req, res) => {
     try {
-        //const incomingData = matchedData(req)
-        const id = req.query.id
-        const activity = req.query.activity
+        const { id, activity } = matchedData(req, { locations: ["query"] })
+        const token = matchedData(req, { locations: ["headers"] }).authorization
+
+        const businessTokenData = verifyToken(token)
+
+        if(id !== businessTokenData._id) throw new Error("ERROR_GET_MAILING_LIST")
 
         const business = await Business.findById(id)
 
         const mailingList = await User.find({
             "city": business.city,
             "interests": activity,
-            "receivesOffers": true  // TODO for some reason recievesOffers is always false even tho some entries have it true
+            "receivesOffers": true
         })        
 
         const mailingListEmails = mailingList.map((user) => {return user.email})
@@ -103,7 +128,7 @@ const getMailingList = async (req, res) => {
         res.send({"emails": mailingListEmails})
     } catch(err) {
         // SLACK log I think
-        handleHTTPError(res, "ERROR_GET_MAILING_LIST")
+        handleHTTPError(res, err.message)
     }
 }
